requirements.txt
cd /Users/parsa_samandi/Documents/Projects/semantic-search-engine

# Edit requirements.txt - remove openai, add sentence-transformers
cat > requirements.txt << 'EOF'
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sentence-transformers==2.2.2
faiss-cpu==1.9.0.post1
numpy==2.2.1
pydantic==2.10.5
python-dotenv==1.0.0
EOF